Auralearn AI Tutor Platform/backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
google-genai>=0.3.0
python-dotenv>=1.0.0,<2.0.0

